backend/requirements.txt
fastapi
uvicorn
python-dotenv
motor
passlib[bcrypt]
python-jose
PyMuPDF
python-docx
pillow
pytesseract
requests
python-multipart

